Split CRUD API to smaller files.
